Performance-related changes.
There was no environment variable set for federalist, so the assets were never minified. Sourcemaps were also removed from the production build, and the build assets file was modified to use external files instead of data uris. Inlining all but the smallest file is a performance anti-pattern, so this provides defaults to account for more likely scenarios. `createAssetPath` was updated to account for the output change necessary to use the external file. assetPaths.json was removed from git because it's a build asset and is generated every time by the build process.